refactor(frontend): group agent studio model selection internals
