+++ packages/frontend/src/pages/agents/model-selection/use-agent-studio-slash-commands.ts
